+++ client/src/containers/AssetAdmin/AssetAdmin.js
@@ -11,7 +11,7 @@ import * as queuedFilesActions from 'state/queuedFiles/QueuedFilesActions';
 import * as displaySearchActions from 'state/displaySearch/DisplaySearchActions';
 import Editor from 'containers/Editor/Editor';
 import Gallery from 'containers/Gallery/Gallery';
-import Breadcrumb from './Breadcrumb';
+import Breadcrumb from './AssetAdminBreadcrumb';
 import Toolbar from 'components/Toolbar/Toolbar';
 import { withApollo } from 'react-apollo';
 import Search, { hasFilters } from 'components/Search/Search';
